:recycle: rework financial document so as not to cause breaking changes
