+++ mindee/documents/financial/__init__.py
@@ -1,2 +1,2 @@
-from .financial_v0 import FinancialV0, TypeFinancialV0
+from .financial_document_v1 import FinancialDocumentV1, TypeFinancialDocumentV1
 from .financial_v1 import FinancialV1, TypeFinancialV1
