+++ tests/test_response.py
@@ -4,7 +4,7 @@ import pytest
 
 from mindee import Client
 from mindee.documents.base import Document
-from mindee.documents.financial.financial_v1 import FinancialV1
+from mindee.documents.financial.financial_document_v1 import FinancialDocumentV1
 from mindee.documents.invoice.invoice_v3 import InvoiceV3
 from mindee.documents.passport.passport_v1 import PassportV1
 from mindee.documents.receipt.receipt_v3 import ReceiptV3
@@ -12,7 +12,9 @@ from mindee.documents.receipt.receipt_v4 import ReceiptV4
 from mindee.endpoints import OTS_OWNER
 from mindee.input.sources import PathInput
 from mindee.response import PredictResponse
-from tests.documents.test_financial_v1 import FILE_PATH_FINANCIAL_DOC_V1_RECEIPT
+from tests.documents.test_financial_document_v1 import (
+    FILE_PATH_FINANCIAL_DOC_V1_RECEIPT,
+)
 from tests.documents.test_invoice_v3 import FILE_PATH_INVOICE_V3_COMPLETE
 from tests.documents.test_passport_v1 import FILE_PATH_PASSPORT_V1_COMPLETE
 from tests.documents.test_receipt_v3 import FILE_PATH_RECEIPT_V3_COMPLETE
@@ -80,15 +82,15 @@ def test_response_receipt_v4(dummy_file_input, dummy_config):
 
 def test_response_financial_doc_with_receipt(dummy_file_input, dummy_config):
     response = json.load(open(FILE_PATH_FINANCIAL_DOC_V1_RECEIPT))
-    parsed_response = PredictResponse[FinancialV1](
-        doc_config=dummy_config[(OTS_OWNER, FinancialV1.__name__)],
+    parsed_response = PredictResponse[FinancialDocumentV1](
+        doc_config=dummy_config[(OTS_OWNER, FinancialDocumentV1.__name__)],
         http_response=response,
         input_source=dummy_file_input,
         response_ok=True,
     )
-    assert isinstance(parsed_response.document, FinancialV1)
+    assert isinstance(parsed_response.document, FinancialDocumentV1)
     for page in parsed_response.pages:
-        assert isinstance(page, FinancialV1)
+        assert isinstance(page, FinancialDocumentV1)
 
 
 def test_response_passport_v1(dummy_file_input, dummy_config):
